modularise compliance runner
