+++ spec/Runners/JasmineComplianceRunner.spec.ts
@@ -1,24 +1,24 @@
-import { JasmineComplianceRunner } from "../../src";
 import { jasmine } from "jasmine";
 import * as AWSMock from 'aws-sdk-mock'
+import { LambdaTestRunner } from "../../src/runners/LambdaTestRunner";
 
 function sleep(ms) {
     return new Promise(resolve => setTimeout(resolve, ms));
-  }
-  
-describe('Given a sample test', ()=> {
-    afterEach(()=> {
+}
+
+describe('Given a sample test', () => {
+    afterEach(() => {
         AWSMock.restore('S3');
     });
-    
-    it('Jasmin compliance runner should execute it successfully', async(done) => {
-        AWSMock.mock("S3", "putObject", (params, callback)=> {
+
+    it('Jasmin compliance runner should execute it successfully', async (done) => {
+        AWSMock.mock("S3", "putObject", (params, callback) => {
             callback(null, "");
         });
         const envStage = 'np'
         var cb = () => { };
         const localExecutionFlag = true;
-        const complianceRunner = new JasmineComplianceRunner(localExecutionFlag);
+        const complianceRunner = new LambdaTestRunner(localExecutionFlag);
         let complianceParams = {};
         require('./sampleTestSuite').suite({});
         await complianceRunner.execute();
@@ -26,8 +26,8 @@ describe('Given a sample test', ()=> {
         //sleep(4000);
         done();
     })
-    it('Jasmin compliance runner should throw an error when it can not upload the report to S3', async(done) => {
-        AWSMock.mock("S3", "putObject", (params, callback)=> {
+    it('Jasmin compliance runner should throw an error when it can not upload the report to S3', async (done) => {
+        AWSMock.mock("S3", "putObject", (params, callback) => {
             callback(new Error('unable to upload to S3'), "");
         });
 
@@ -35,12 +35,11 @@ describe('Given a sample test', ()=> {
         var cb = () => { };
         const localExecutionFlag = true;
         try {
-            const complianceRunner = new JasmineComplianceRunner(localExecutionFlag);
+            const complianceRunner = new LambdaTestRunner(localExecutionFlag);
             require('./sampleTestSuite').suite({});
             await complianceRunner.execute();
             await complianceRunner.uploadJUnitReportToS3('MyProduct', 'MyBucket');
-        } catch (error){
-            console.log(`error: ${error}`)
+        } catch (error) {
             done();
             return;
         }
